requirements.txt
# AI夜宵爆品预测助手依赖

# Web框架
fastapi==0.115.6
uvicorn==0.34.0

# 数据库
sqlalchemy==2.0.36
pymysql==1.1.1
redis==5.0.1

# 数据验证
pydantic==2.10.4
pydantic-settings==2.7.1

# 认证
python-jose==3.3.0
passlib==1.7.4

# 异步任务
celery==5.4.0

# 机器学习/时序预测
prophet==1.1.5
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.8.0

# LLM
langchain==1.2.6

# HTTP客户端
httpx==0.28.1

# 工具
python-multipart==0.0.6
PyYAML==6.0.3
python-json-logger==2.0.7
